+++ EarTrumpet/DataModel/WindowsAudio/Internal/AudioDeviceSession.cs
@@ -10,6 +10,7 @@ using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Threading;
 using Windows.Win32;
 using Windows.Win32.Media.Audio;
@@ -33,29 +34,32 @@ internal class AudioDeviceSession : BindableBase, IAudioSessionEvents, IAudioDev
 
     public float Volume
     {
-        get => App.Settings.UseLogarithmicVolume ? _volume.ToDisplayVolume() : _volume;
+        get => App.Settings.UseLogarithmicVolume
+            ? _volume.LinearToLog()
+            : _volume;
         set
         {
-            value = value.Bound(0, 1f);
-
-            if (App.Settings.UseLogarithmicVolume)
+            try
             {
-                value = value.ToLogVolume();
-            }
-
-            if (_volume != value)
-            {
-                try
+                if (App.Settings.UseLogarithmicVolume)
                 {
+                    value = value.Bound(App.Settings.LogarithmicVolumeMinDb, 0);
+                    // We must convert manually here because sessions use linear volume.
+                    _simpleVolume.SetMasterVolume(value.LogToLinear(), Guid.Empty);
                     _volume = value;
-                    _simpleVolume.SetMasterVolume(value, Guid.Empty);
+                    IsMuted = value <= App.Settings.LogarithmicVolumeMinDb;
                 }
-                catch (Exception ex) when (ex.Is(HRESULT.AUDCLNT_E_DEVICE_INVALIDATED))
+                else
                 {
-                    // Expected in some cases.
+                    value = value.Bound(0, 1f);
+                    _simpleVolume.SetMasterVolume(value, Guid.Empty);
+                    _volume = value;
+                    IsMuted = _volume.ToVolumeInt() == 0;
                 }
-
-                IsMuted = App.Settings.UseLogarithmicVolume ? _volume <= (1 / 100f).ToLogVolume() : _volume.ToVolumeInt() == 0;
+            }
+            catch (Exception ex) when (ex.Is(HRESULT.AUDCLNT_E_DEVICE_INVALIDATED))
+            {
+                // Expected in some cases.
             }
         }
     }
@@ -189,6 +193,16 @@ internal class AudioDeviceSession : BindableBase, IAudioSessionEvents, IAudioDev
                 SyncPersistedEndpoint(parent);
             }
         }
+
+        WeakEventManager<AppSettings, EventArgs>.AddHandler(
+            App.Settings,
+            nameof(AppSettings.UseLogarithmicVolumeChanged),
+            UseLogarithmicVolumeChangedHandler);
+    }
+
+    private void UseLogarithmicVolumeChangedHandler(object sender, EventArgs e)
+    {
+        RaisePropertyChanged(nameof(Volume));
     }
 
     ~AudioDeviceSession()
@@ -242,8 +256,16 @@ internal class AudioDeviceSession : BindableBase, IAudioSessionEvents, IAudioDev
     public void UpdatePeakValueBackground()
     {
         var newValues = Helpers.ReadPeakValues(_meter);
-        PeakValue1 = newValues[0];
-        PeakValue2 = newValues[1];
+        if (App.Settings.UseLogarithmicVolume)
+        {
+            PeakValue1 = newValues[0].LinearToLogNormalized();
+            PeakValue2 = newValues[1].LinearToLogNormalized();
+        }
+        else
+        {
+            PeakValue1 = newValues[0];
+            PeakValue2 = newValues[1];
+        }
     }
 
     private void ChooseDisplayName(string displayNameFromSession)